Dominant Embeddin Index/Dominant Embeddin Index/bulid Index/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1
_openmp_mutex=4.5
absl-py=1.1.0
asttokens=2.0.5
backcall=0.2.0
backports=1.0
backports.functools_lru_cache=1.6.4
blas=1.0
brotlipy=0.7.0
bzip2=1.0.8
ca-certificates=2022.5.18.1
cachetools=5.2.0
certifi=2022.5.18.1
cffi=1.15.0
charset-normalizer=2.0.12
class-resolver=0.3.10
cryptography=37.0.2
cudatoolkit=11.3.1
cycler=0.11.0
debugpy=1.6.0
decorator=5.1.1
deepsnap=0.2.1
entrypoints=0.4
executing=0.8.3
ffmpeg=4.3
freetype=2.10.4
future=0.18.2
fvcore=0.1.5.post20221221
gmp=6.2.1
gnutls=3.6.13
google-auth=2.8.0
google-auth-oauthlib=0.4.6
grpcio=1.47.0
icu=67.1
idna=3.3
imageio=2.19.3
importlib-metadata=4.11.4
intel-openmp=2021.4.0
iopath=0.1.10
ipykernel=6.13.1
ipython=8.4.0
jedi=0.18.1
jinja2=3.1.2
joblib=1.1.0
jpeg=9e
jupyter_client=7.3.4
jupyter_core=4.10.0
kiwisolver=1.4.2
lame=3.100
ld_impl_linux-64=2.36.1
libblas=3.9.0
libcblas=3.9.0
libffi=3.4.2
libgcc-ng=12.1.0
libgfortran-ng=12.1.0
libgfortran5=12.1.0
libgomp=12.1.0
libiconv=1.14
liblapack=3.9.0
libnsl=2.0.0
libpng=1.6.37
libsodium=1.0.18
libstdcxx-ng=12.1.0
libtiff=4.0.10
libuuid=2.32.1
libuv=1.43.0
libzlib=1.2.12
lz4-c=1.9.3
markdown=3.3.7
markupsafe=2.1.1
matplotlib=3.2.2
matplotlib-base=3.2.2
matplotlib-inline=0.1.3
metis=0.2a5
mkl=2021.4.0
mkl-service=2.4.0
mkl_fft=1.3.1
mkl_random=1.2.2
ncurses=6.3
nest-asyncio=1.5.5
nettle=3.6
networkit=10.1
networkx=2.8.3
numpy=1.22.3
numpy-base=1.22.3
oauthlib=3.2.0
olefile=0.46
openh264=2.1.1
openssl=1.1.1o
packaging=21.3
pandas=1.4.2
parso=0.8.3
pexpect=4.8.0
pickleshare=0.7.5
pillow=9.1.1
pip=22.1.2
portalocker=2.7.0
prompt-toolkit=3.0.29
protobuf=3.19.4
psutil=5.9.1
ptyprocess=0.7.0
pure_eval=0.2.2
pyasn1=0.4.8
pyasn1-modules=0.2.8
pycparser=2.21
pyg=2.0.4
pygments=2.12.0
pymetis=2023.1
pyopenssl=22.0.0
pyparsing=3.0.9
pysocks=1.7.1
python=3.8.13
python-dateutil=2.8.2
python-louvain=0.15
python_abi=3.8
pytorch=1.11.0
pytorch-cluster=1.6.0
pytorch-mutex=1.0
pytorch-scatter=2.0.9
pytorch-sparse=0.6.13
pytorch-spline-conv=1.2.1
pytz=2022.1
pyyaml=6.0
pyzmq=23.1.0
readline=8.1.2
requests=2.28.0
requests-oauthlib=1.3.1
rsa=4.8
rtree=1.0.0
scikit-learn=1.1.1
scipy=1.8.1
seaborn=0.12.1
setuptools=62.3.4
six=1.16.0
sqlite=3.38.5
stack_data=0.2.0
tabulate=0.9.0
tensorboard=2.9.1
tensorboard-data-server=0.6.1
tensorboard-plugin-wit=1.8.1
termcolor=2.2.0
test-tube=0.7.5
threadpoolctl=3.1.0
tk=8.6.12
torchaudio=0.11.0
torchvision=0.12.0
tornado=6.1
tqdm=4.56.0
traitlets=5.2.2.post1
typing_extensions
urllib3=1.26.9
wcwidth=0.2.5
werkzeug=2.1.2
wheel=0.37.1
xz=5.2.5
yacs=0.1.8
yaml=0.2.5
zeromq=4.3.4
zipp=3.8.0
zlib=1.2.12
zstd=1.4.9
